refactor: StaffOut/InPageを共通コンポーネントに統合
